inspection-ai/requirements.txt
streamlit==1.31.0
snowflake-snowpark-python==1.12.0
pandas==2.1.4
plotly==5.18.0

pillow==10.2.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pypdf==4.0.1
